+++ src/main/resources/application.properties
@@ -26,3 +26,4 @@ spring.jpa.hibernate.ddl-auto =update
 spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.MySQL5Dialect
 
 spring.messages.basename=validation
+server.port=8081
